requirements.txt
fastapi==0.115.6
pydantic==2.10.3
uvicorn==0.32.1
pymongo==4.10.1
python-dotenv==1.0.1
pyaml-env==1.2.1
PyYAML==6.0.1
pytest==7.4.4
pytest-asyncio==0.23.5
mongomock==4.1.2
httpx==0.26.0
setuptools==70.0.0
phonenumbers==8.13.51
validators~=0.34.0
python-multipart==0.0.6